python-data-engineer/week12/exercises/index_optimization.py
"""
Index optimization exercises.
"""

from typing import Dict, Any, Optional, List, Union
import logging
from abc import ABC, abstractmethod
from pyspark.sql import SparkSession
from pyspark.sql.functions import *
from pyspark.sql.types import *

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class Index(ABC):
    """Abstract base class for indexes."""
    
    @abstractmethod
    def create(self, table_name: str, columns: List[str]) -> None:
        """Create index."""
        pass
    
    @abstractmethod
    def drop(self, table_name: str) -> None:
        """Drop index."""
        pass
    
    @abstractmethod
    def get_stats(self) -> Dict[str, Any]:
        """Get index statistics."""
        pass

class BTreeIndex(Index):
    """B-tree index implementation."""
    
    def __init__(self, spark: SparkSession):
        """Initialize B-tree index."""
        self.spark = spark
        self.stats = {}
    
    def create(self, table_name: str, columns: List[str]) -> None:
        """
        Create B-tree index.
        
        Args:
            table_name: Table name
            columns: Columns to index
        """
        try:
            # Create index using Delta Lake Z-ORDER
            self.spark.sql(f"""
                OPTIMIZE {table_name}
                ZORDER BY ({','.join(columns)})
            """)
            
            # Store statistics
            self.stats[table_name] = {
                'type': 'btree',
                'columns': columns,
                'created_at': current_timestamp()
            }
            
            logger.info(f"Created B-tree index on {table_name}")
            
        except Exception as e:
            logger.error(f"Failed to create B-tree index: {e}")
            raise
    
    def drop(self, table_name: str) -> None:
        """
        Drop B-tree index.
        
        Args:
            table_name: Table name
        """
        try:
            # Remove statistics
            self.stats.pop(table_name, None)
            
            logger.info(f"Dropped B-tree index from {table_name}")
            
        except Exception as e:
            logger.error(f"Failed to drop B-tree index: {e}")
            raise
    
    def get_stats(self) -> Dict[str, Any]:
        """Get index statistics."""
        return self.stats

class BitmapIndex(Index):
    """Bitmap index implementation."""
    
    def __init__(self, spark: SparkSession):
        """Initialize bitmap index."""
        self.spark = spark
        self.stats = {}
    
    def create(self, table_name: str, columns: List[str]) -> None:
        """
        Create bitmap index.
        
        Args:
            table_name: Table name
            columns: Columns to index
        """
        try:
            # Create bitmap index table
            bitmap_table = f"{table_name}_bitmap"
            
            for column in columns:
                self.spark.sql(f"""
                    CREATE TABLE IF NOT EXISTS {bitmap_table}_{column}
                    USING DELTA
                    AS SELECT 
                        {column},
                        collect_set(id) as row_ids
                    FROM {table_name}
                    GROUP BY {column}
                """)
            
            # Store statistics
            self.stats[table_name] = {
                'type': 'bitmap',
                'columns': columns,
                'created_at': current_timestamp()
            }
            
            logger.info(f"Created bitmap index on {table_name}")
            
        except Exception as e:
            logger.error(f"Failed to create bitmap index: {e}")
            raise
    
    def drop(self, table_name: str) -> None:
        """
        Drop bitmap index.
        
        Args:
            table_name: Table name
        """
        try:
            # Drop bitmap tables
            bitmap_table = f"{table_name}_bitmap"
            
            for column in self.stats[table_name]['columns']:
                self.spark.sql(f"""
                    DROP TABLE IF EXISTS {bitmap_table}_{column}
                """)
            
            # Remove statistics
            self.stats.pop(table_name, None)
            
            logger.info(f"Dropped bitmap index from {table_name}")
            
        except Exception as e:
            logger.error(f"Failed to drop bitmap index: {e}")
            raise
    
    def get_stats(self) -> Dict[str, Any]:
        """Get index statistics."""
        return self.stats

class HashIndex(Index):
    """Hash index implementation."""
    
    def __init__(self, spark: SparkSession):
        """Initialize hash index."""
        self.spark = spark
        self.stats = {}
    
    def create(self, table_name: str, columns: List[str]) -> None:
        """
        Create hash index.
        
        Args:
            table_name: Table name
            columns: Columns to index
        """
        try:
            # Create hash index table
            hash_table = f"{table_name}_hash"
            
            self.spark.sql(f"""
                CREATE TABLE IF NOT EXISTS {hash_table}
                USING DELTA
                AS SELECT 
                    hash({','.join(columns)}) as hash_key,
                    collect_set(id) as row_ids
                FROM {table_name}
                GROUP BY hash({','.join(columns)})
            """)
            
            # Store statistics
            self.stats[table_name] = {
                'type': 'hash',
                'columns': columns,
                'created_at': current_timestamp()
            }
            
            logger.info(f"Created hash index on {table_name}")
            
        except Exception as e:
            logger.error(f"Failed to create hash index: {e}")
            raise
    
    def drop(self, table_name: str) -> None:
        """
        Drop hash index.
        
        Args:
            table_name: Table name
        """
        try:
            # Drop hash table
            hash_table = f"{table_name}_hash"
            self.spark.sql(f"""
                DROP TABLE IF EXISTS {hash_table}
            """)
            
            # Remove statistics
            self.stats.pop(table_name, None)
            
            logger.info(f"Dropped hash index from {table_name}")
            
        except Exception as e:
            logger.error(f"Failed to drop hash index: {e}")
            raise
    
    def get_stats(self) -> Dict[str, Any]:
        """Get index statistics."""
        return self.stats

class IndexOptimizer:
    """Index optimization management."""
    
    def __init__(self, spark: SparkSession):
        """Initialize optimizer."""
        self.spark = spark
        self.indexes = {
            'btree': BTreeIndex(spark),
            'bitmap': BitmapIndex(spark),
            'hash': HashIndex(spark)
        }
    
    def suggest_indexes(
        self,
        table_name: str,
        query_patterns: List[Dict[str, Any]]
    ) -> List[Dict[str, Any]]:
        """
        Suggest indexes based on query patterns.
        
        Args:
            table_name: Table name
            query_patterns: List of query patterns
        
        Returns:
            List of index suggestions
        """
        try:
            suggestions = []
            
            for pattern in query_patterns:
                if pattern.get('equality_predicates'):
                    suggestions.append({
                        'type': 'hash',
                        'columns': pattern['equality_predicates'],
                        'reason': 'Equality predicates benefit from hash indexes'
                    })
                
                if pattern.get('range_predicates'):
                    suggestions.append({
                        'type': 'btree',
                        'columns': pattern['range_predicates'],
                        'reason': 'Range predicates benefit from B-tree indexes'
                    })
                
                if pattern.get('low_cardinality_columns'):
                    suggestions.append({
                        'type': 'bitmap',
                        'columns': pattern['low_cardinality_columns'],
                        'reason': 'Low cardinality columns benefit from bitmap indexes'
                    })
            
            return suggestions
            
        except Exception as e:
            logger.error(f"Failed to suggest indexes: {e}")
            raise
    
    def create_index(
        self,
        index_type: str,
        table_name: str,
        columns: List[str]
    ) -> None:
        """
        Create index.
        
        Args:
            index_type: Type of index
            table_name: Table name
            columns: Columns to index
        """
        try:
            if index_type not in self.indexes:
                raise ValueError(f"Invalid index type: {index_type}")
            
            self.indexes[index_type].create(table_name, columns)
            
        except Exception as e:
            logger.error(f"Failed to create index: {e}")
            raise
    
    def drop_index(
        self,
        index_type: str,
        table_name: str
    ) -> None:
        """
        Drop index.
        
        Args:
            index_type: Type of index
            table_name: Table name
        """
        try:
            if index_type not in self.indexes:
                raise ValueError(f"Invalid index type: {index_type}")
            
            self.indexes[index_type].drop(table_name)
            
        except Exception as e:
            logger.error(f"Failed to drop index: {e}")
            raise
    
    def get_index_stats(
        self,
        index_type: Optional[str] = None
    ) -> Dict[str, Any]:
        """
        Get index statistics.
        
        Args:
            index_type: Type of index (optional)
        
        Returns:
            Index statistics
        """
        try:
            if index_type:
                if index_type not in self.indexes:
                    raise ValueError(f"Invalid index type: {index_type}")
                return self.indexes[index_type].get_stats()
            
            return {
                type_: index.get_stats()
                for type_, index in self.indexes.items()
            }
            
        except Exception as e:
            logger.error(f"Failed to get index stats: {e}")
            raise 